feat(events): created event and event category schemas; wired all event routes
