+++ backend/app/main.py
@@ -6,9 +6,11 @@ from scalar_fastapi import get_scalar_api_reference
 from app.core.config import get_settings
 from app.core.security import verify_api_key
 from app.db.init_db import init_db, seed_dev_data
-from app.api.routes import tournaments, sheets, events, users
-from app.api.routes import auth
-from backend.app.api.routes import tournament_memberships
+from app.api.routes import (
+    auth, events, tournaments,
+    tournament_events, tournament_memberships,
+    sheets, users
+)
 
 settings = get_settings()
 
@@ -50,12 +52,13 @@ api_key_dependency = Depends(verify_api_key)
 
 # All routes require API key — including auth (login, logout, register).
 # In development with API_KEY unset, security.py skips the check automatically.
-app.include_router(auth.router,        prefix="", dependencies=[api_key_dependency])
-app.include_router(tournaments.router, prefix="", dependencies=[api_key_dependency])
-app.include_router(events.router,      prefix="", dependencies=[api_key_dependency])
+app.include_router(auth.router,                   prefix="", dependencies=[api_key_dependency])
+app.include_router(events.router,                 prefix="", dependencies=[api_key_dependency])
+app.include_router(tournaments.router,            prefix="", dependencies=[api_key_dependency])
+app.include_router(tournament_events.router,      prefix="", dependencies=[api_key_dependency])
 app.include_router(tournament_memberships.router, prefix="", dependencies=[api_key_dependency])
-app.include_router(sheets.router,      prefix="", dependencies=[api_key_dependency])
-app.include_router(users.router,       prefix="", dependencies=[api_key_dependency])
+app.include_router(sheets.router,                 prefix="", dependencies=[api_key_dependency])
+app.include_router(users.router,                  prefix="", dependencies=[api_key_dependency])
 
 
 @app.get("/health", tags=["meta"])
